Move redacted tests into their own file
